Introduce $posted_data_hash and $skip_spam_check properties
Generate a hash based on `$posted_data` after spam verification. Having a hash identical to one from a past submission means the current submission is identical to the past one and it has once passed through a spam verification (has confirmed it is not spam), so you can skip spam check safely this time. This avoids to apply unnecessary spam checks to submissions that are safe in reality.

See #57