Update config-validator.php
Hard message limit is good for most of the installations. But some of the servers accept even 100 megabytes messages. In other cases the sum of all limits in field type inputs will never be reached cause some of the fields show conditionally. That's why it would be good to give user control over this limit. This update will add a new filter allowing devs to change the limit according to their needs with keeping the default limit to 25 megabytes.