Enable async loading of google recaptcha
Hi, I don't know if this is a very elegant solution to my problem, but I will present it nevertheless. :)

I was trying to load the google-recaptcha script async and and also integrate it into our cookie consent tool (usercentics cmp v2).

To integrate it into the cmp the type has to be set to "text/plain" and additionally the script needs a data attribute ( data-usercentrics='reCaptcha v3' ), 
this is done with the filter 'script_loader_tag'.

If the user gives consent to the specific service the cmp will enable all the code and run it.
As cf7 attaches itself to DOMContentLoaded it never fired for this scenario.
index.js is also set so async and text/plain on load, 

!!! I don't know what happens with the wait timer when grecaptcha never comes online. (which does not happen in my scenario)

I think there is additional work needed to get it nice and clean, but it works for me now. Thank you.